+++ includes/pu-visitor-log-admin-page.php
@@ -26,8 +26,15 @@ class pu_visitor_log_admin_page
     {
         echo '<div class="wrap">';
         echo '<h2>Your Plugin Name</h2>';
-        echo 'test ';
+        // Display whatever it is you want to show.
+        echo 'Geef hier de mogelijkheden aan die u graag voor de leden van uw ondernemersverening beschikbaar wilt hebben.';
+        echo '<form id="pu_visitor_settings" action="options.php" method="post">';
+        settings_fields('pu_visitor_settings');
+        do_settings_sections('pu-visitor-log');
+        submit_button('Opslaan', 'primary', 'ng_ondernemersvereniging_submit');
+        echo '</form>';
         echo '</div>';
     }
 
 }
+