+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # [wonder32](https://www.newfountain.nl)
 
 
+## 0.0.4 settings (26 july, 2016)
+* working settings group
+* register setting linked to settings group
+
+## 0.0.3 (de)activation (24 july, 2016)
+* (de)activation class
+* register (de)activation
+* save some options
+
 ## 0.0.2 admin page (23 july, 2016)
 
 * update functionality